Add header comments and specify imports
