+++ src/main/java/langley/tasklist/TaskList.java
@@ -15,22 +15,45 @@ public class TaskList {
         return tasks.get(index);
     }
 
+    /**
+     * Adds a task to the task list.
+     *
+     * @param task Task to be added to the list.
+     */
     public void addTask(Task task) {
         tasks.add(task);
     }
 
+    /**
+     * Deletes a task from the task list by its index.
+     *
+     * @param index Index of the task to be deleted.
+     */
     public void deleteTask(int index) {
         tasks.remove(index);
     }
 
+    /**
+     * Marks the task at the specified index as completed.
+     *
+     * @param index Index of the task to be marked.
+     */
     public void markTask(int index) {
         tasks.get(index).mark();
     }
 
+    /**
+     * Unmarks the task at the specified index, marking it as incomplete.
+     *
+     * @param index Index of the task to be unmarked.
+     */
     public void unmarkTask(int index) {
         tasks.get(index).unmark();
     }
 
+    /**
+     * Lists all tasks currently in the task list, printing them with their index and status.
+     */
     public void listTasks() {
         for (int i = 0; i < tasks.size(); i++) {
             System.out.println((i + 1) + ". " + tasks.get(i));
